server/Job Schedular/requirements.txt
requests == 2.25.1
email-to == 0.1.0
secure-smtplib == 0.1.1
pymongo == 3.12.1
dnspython == 2.1.0
python-dotenv == 0.19.2